fix story page tests
